Better login-related redirects
If the user visits the app on a page other than the landing or a page related to logging in, they're redirected to the login page instead of the landing. Upon seeing that page, they're shown an alert saying that they need to log in to see that page. Logging in in this state will, following a successful login, send the user to the page that they originally navigated to (via external link or bookmark).